Quick Save

+++ app/src/main/java/com/krpvartstudio/sshine/business/model/MainHourListModel.kt
@@ -3,10 +3,10 @@ package com.krpvartstudio.sshine.business.model
 data class MainHourListModel(
     val clouds: Int,
     val dew_point: Double,
-    val dt: Int,
+    val dt: Long,
     val feels_like: Double,
     val humidity: Int,
-    val pop: Int,
+    val pop: Double,
     val pressure: Int,
     val temp: Double,
     val uvi: Int,
